chore: reorganize selenium tests
- rename `ui-selenium` task as `ui`
- rename `test/selenium` to `test/ui`
- move tests out of formed `ui` folder which was redundant
- add numeric prefix to test files to control their execution order
  based on importance (we still expect to be able to run any test
  in separation without having to run others before it).
