+++ docker-compose.yml
@@ -1,4 +1,4 @@
-# Used by `task ui-selenium` to start the selenium server.
+# Used by `task ui` to start the selenium server.
 
 # Override VSIX path: VSX_FILE=./out/ansible-1.2.3.vsix
 # cspell: ignore healthcheck
@@ -15,7 +15,7 @@ services:
       # apparently podman does not want to mount file as container root
       - ${VSX_FILE:-./out/data/ansible-latest.vsix}:/data/ansible-latest.vsix:ro
       - ./.vscode-test/user-data/User/settings.json:/home/selenium/.local/share/code-server/User/settings.json:rw
-      - ./out/ui-selenium/logs:/home/selenium/.local/share/code-server/logs:rw
+      - ./out/ui/logs:/home/selenium/.local/share/code-server/logs:rw
     restart: "no"
     healthcheck:
       test: ["CMD", "curl", "-f", "http://localhost:4444"]
