--- test/selenium/ui/__init__.py
@@ -1,1 +1,0 @@
-"""UI tests for VSCode Ansible extension using Selenium."""
