+++ test/ui/hooks/logging_hook.py
