+++ test/ui/test_02_welcome_webviews.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import pytest
 
-from test.selenium.utils.ui_utils import (
+from test.ui.utils.ui_utils import (
     ensure_vscode_ready,
     find_element_across_iframes,
     vscode_run_command,
