+++ test/ui/test_80_lightspeed.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import pytest
 
-from test.selenium.utils.ui_utils import (
+from test.ui.utils.ui_utils import (
     find_element_across_iframes,
     vscode_explanation,
     vscode_login,
